Arraged the files and quad is pretty working littler with no errors
